Added 1 PSP multiplayer
So a lot of code changed... hers a summary:
Added code for multiplayer on 1 PSP
Changed code in mode == 1 to work with this MP without needing to
recreate the "game"
added exception to initGame() to stop random gen for the colors when set
by P1
Added clearGC() function to clear all GeneratedCode so P1 can change it
from a fresh every time.

Other:
Added GameOver/Win screens instead of text
Pins are empty instead of light green if not yet changed.
PinX,PinY and image loading moved to render.lua